requirements.txt
weaviate-client==3.15.1
sentence-transformers==4.0.2
numpy==2.2.4
pandas==2.2.3
networkx==3.4.2
matplotlib==3.10.1
torch==2.6.0
scikit-learn==1.6.1
transformers==4.51.2
pytest==8.3.5
langgraph==0.3.31
langchain-openai==0.3.14
langchain-community==0.3.21
openai==1.73.0
python-dotenv==1.1.0
faiss-cpu==1.10.0
tqdm==4.67.1
rouge-score==0.1.2